+++ Dockerfile
@@ -11,7 +11,7 @@ FROM scratch
 WORKDIR /
 
 COPY ./bin/everest  /everest-api
-COPY ./bin/helm-tools/ /helm-tools
+COPY ./bin/everest-helm-tools/ /everest-helm-tools
 COPY --from=build /etc/ssl/certs/ca-certificates.crt /etc/ssl/certs/
 
 EXPOSE 8080
