+++ Makefile
@@ -20,13 +20,13 @@ help:                   ## Display this help message
 init:                   ## Install development tools
 	cd tools && go generate -x -tags=tools
 
-build-helm-tools: 		## Build binaries for Helm chart hooks.
-	cd helm-tools && go generate -x
+build-helm-tools-cli: 			## Build the CLI for helm-tools
+	cd helm && go generate -x
 
 build: build-helm-tools 				## Build binaries 
 	go build -v $(LD_FLAGS_API) -o bin/everest ./cmd
 
-build-cli:        		## Build binaries
+build-cli: build-helm-tools-cli        	## Build binaries 
 	go build -tags debug -v $(LD_FLAGS_CLI_TEST) -o bin/everestctl ./cmd/cli
 
 release: FLAGS += -X 'github.com/percona/everest/cmd/config.TelemetryURL=https://check.percona.com' -X 'github.com/percona/everest/cmd/config.TelemetryInterval=24h'
